Remove last warnings

+++ botcraft/src/Game/World/Blockstate.cpp
@@ -720,7 +720,7 @@ namespace Botcraft
 
         speed_multiplier += (tool_efficiency > 0) * 1.0f + tool_efficiency * tool_efficiency;
         speed_multiplier *= 1.0f + 0.2f * haste;
-        speed_multiplier *= std::pow(0.3f, std::min(static_cast<int>(fatigue), 4));
+        speed_multiplier *= static_cast<float>(std::pow(0.3f, std::min(static_cast<int>(fatigue), 4)));
         speed_multiplier *= head_in_fluid_wo_aqua_affinity ? 0.2f : 1.0f;
         speed_multiplier *= on_ground ? 1.0f : 0.2f;
